backend/main_api/requirements.txt
fastapi
uvicorn
google-adk==1.5.0
httpx
a2a-sdk==0.2.10
python-dotenv
asyncclick
pytest
click